+++ src/traits.rs
@@ -1,0 +1,3 @@
+pub trait BackedType {
+    fn get_conn(&mut self) -> &mut redis::Connection;
+}